Python Programs - CSE 110
Adventure Game Milestone - It's a little game that was proposed to create choices, consequences, and a little story. So depending on your choice, you'll be going through a different situation, either good or bad.
Area of a Shape - It's a calculator that calculates the area, length, side, and radius of a shaped rectangle, square and circle
Clever Stories - It's a game for you to create a funny story using adjectives, animals, verbs, and exclamation, and continue the story as you want.
Comparing - It's a program that gets the user input number, says if the number is greater or equal, and asks for your favorite animal.
Converting Temperatures - It's a calculator that converts Fahrenheit to Celcius
Data Analysis - This Program reads a CSV file that contains countries of the world, their code, the year, and life expectancy in years. Based on that, my program gets user input according to the year it wants to compare, so when the program runs it will show the the countries with highest and lowest life expectancy based on the year the user typed.
Grade Calculator - It will tell the user, according to its input, if the grade is A, B, C, D, or  F minus or plus as well
Human Resources - This is another project that reads a file, this time a TXT file, that contains the name, ID, profession, and salary.
ID Card - Will print an ID Card according to the user's input
Items List - You can create a list of items and you can decide to change an item, add a new item, and check your list
Meal Price Calculator - It will calculate the amount you need to pay for the meal and the waiter's tip to pay if you want
Qualifying for Loan - According to the user's information the program will tell if the user is eligible to receive a loan or not from the "bank" 
Secret Word - This one is a trick you need to guess the word of another person, but since you play it alone, you will know the secret word always
Shopping Cart Milestone - This is a shipping cart list that you can see what is in it, you can change the item, add a new item, put the price of the products, and check your cart whenever you want
Wildchill Calculator - The Windchill calculator will show you how high is the Temperature according to the speed of the wind, So the fast you moves the height is the temperature in Fahrenheit or Celcius